requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
av==17.0.0
bitsandbytes==0.49.2
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
datasets==4.7.0
dill==0.4.0
exceptiongroup==1.3.1
filelock==3.25.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
narwhals==2.18.1
networkx==3.4.2
numpy==2.2.6
packaging==26.0
pandas==2.3.3
peft==0.18.1
pillow==12.1.1
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
pyarrow==23.0.1
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-pptx==1.0.2
pytz==2026.1.post1
PyYAML==6.0.3
qwen-vl-utils==0.0.14
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
torchvision==0.25.0
tornado==6.5.5
tqdm==4.67.3
transformers==5.3.0
trl==0.29.0
typer==0.24.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.23.0
